apps/studio/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# File handling
aiofiles==23.2.0

# Image processing
Pillow==10.1.0

# Environment variables
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Background tasks and queuing
celery==5.3.4
redis==5.0.1

# Database and migrations
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1
# asyncpg==0.29.0  # PostgreSQL

# Cloud storage (optional)
# supabase==2.0.2
# boto3==1.34.0  # AWS S3

# AI/ML libraries
# torch==2.1.1  # For local model inference
# transformers==4.36.0  # HuggingFace models
# diffusers==0.24.0  # Diffusion models

# Security
cryptography==44.0.3
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI

# Code quality (development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
